+++ includes/utils.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Utility functions
+ *
+ * @since  1.7
+ * @package  10up-experience
+ */
+
+namespace TenUpExperience\Utils;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit; // Exit if accessed directly.
+}
+
+
+/**
+ * Whether plugin is network activated
+ *
+ * Determines whether plugin is network activated or just on the local site.
+ *
+ * @since 1.7
+ * @param string $plugin the plugin base name.
+ * @return bool True if network activated or false.
+ */
+function is_network_activated( $plugin ) {
+	$plugins = get_site_option( 'active_sitewide_plugins' );
+
+	if ( is_multisite() && isset( $plugins[ $plugin ] ) ) {
+		return true;
+	}
+
+	return false;
+}